+++ libs/coin-polkadot/.unimportedrc.json
@@ -3,11 +3,6 @@
     "src/bridge/index.ts",
     "src/bridge/deviceTransactionConfig.ts",
     "src/bridge/formatters.ts",
-    "src/bridge/hw-getAddress.ts",
-    "src/bridge/logic.ts",
-    "src/bridge/preload.ts",
-    "src/bridge/serialization.ts",
-    "src/bridge/transaction.ts",
     "src/signer/index.ts",
     "src/test/cli.ts",
     "src/test/index.ts",
