+++ libs/coin-polkadot/src/logic/estimateMaxSpendable.ts
@@ -2,7 +2,7 @@ import { BigNumber } from "bignumber.js";
 import type { AccountLike, Account } from "@ledgerhq/types-live";
 import { getMainAccount } from "@ledgerhq/coin-framework/account/index";
 import type { PolkadotAccount, Transaction } from "../types";
-import { calculateAmount } from "./logic";
+import { calculateAmount } from "./utils";
 import getEstimatedFees from "./getFeesForTransaction";
 import { createTransaction } from "./transaction";
 import { loadPolkadotCrypto } from "./polkadot-crypto";
