+++ libs/coin-polkadot/src/signer/index.ts
@@ -2,6 +2,6 @@
  * This directory is the home for all types and logic based on Ledgers signer.
  */
 
-import resolver from "./hw-getAddress";
+import getAddress from "./getAddress";
 
-export default resolver;
+export default getAddress;
