+++ libs/coin-polkadot/src/test/index.ts
@@ -2,5 +2,5 @@ import makeCliTools from "./cli";
 
 export * from "./bridge.integration.test";
 export { makeCliTools };
-export * from "./specs";
-export * from "./speculos-deviceActions";
+export * from "./bot-specs";
+export * from "./bot-deviceActions";
