+++ libs/ledger-live-common/src/families/polkadot/bridge.integration.test.ts
@@ -1,7 +1,5 @@
 import "../../__tests__/test-helpers/setup";
 import { testBridge } from "../../__tests__/test-helpers/bridge";
-import { dataset } from "@ledgerhq/coin-polkadot/test/bridge.integration.test";
-
-import "@ledgerhq/coin-polkadot/test/bridge.integration.test";
+import { dataset } from "@ledgerhq/coin-polkadot/test/index";
 
 testBridge(dataset);
